requirements.txt
# TO install dependencies run 
# "pip3.8 install -r requirements.txt"
matplotlib>=3.1.2
numpy>=1.18.1
pandas>=0.25.3
progressbar2>=3.47.0
pymc3==3.8
scipy>=1.4.1